requirements.txt
Flask==3.1.0
google-generativeai==0.8.4
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0
waitress==3.0.0
